Join wrapped lines
When a line is more than 998 char long, Bacula wraps it (and loose one
char).  Previously we only parsed the first line and everything left
after the trucation was lost.  This change attempts to gather the rest
of the string.  A char is lost at each wrapping position due to the bug
in Bacula.