Add parameters and make them mandatory
Bacula logging is quite fragile, and lines are trucated at circa 5000
chars, which cause all sorts of trouble when a lot of volumes are used.
Add parameters to pass all base information about backups as parameters
and rely on these information to generate base metrics.  Still process
stdin for more metrics but avoid generating events when we are not 100%
sure they not tructated.