Fix how we set the host of bacula events
In #7, we add parameters to more reliably generate bacula events.  The
`--client` parameter value is overridden by the `--event-host` parameter
when it is set, so remove the --client one and rely on riemann-tools
base arguments to set the host.

Because #7 was never released, this is not considered a breaking change.