Add actual status to bacula status events
When a backup output is truncated, `data['Termination']` is not
available and the status report description might be an empty string.
This is not an issue if the state is 'ok' (corresponding to a
Termination of 'Backup OK') but as any other status result in a
'critical' state, we do not have a feedback on the actual error when
this happen with a trucated message.

Add the status passed as a parameter in the description so that we can
have relevant feedback on error.