Do not update canceled backups state
When a backup is canceled, we do not want to replace the previous state
because we do not know if this backup is cancanceled because terrible
things are happening and the administrator does it best to avoid more
damage, or because the scheduled backup does not make sense (e.g. wrong
host) and the previous backup is perfectly fine.  In any case, the
metric will eventualy timeout or be replaced with an actual new state at
some point in the future.